requirements.txt
streamlit==1.32.0
Pillow>=10.0.0
requests==2.31.0
mysql-connector-python==8.3.0
python-dotenv==1.0.1
bcrypt==4.1.2
numpy>=1.19.3,<2.0.0